pyproject.toml
[tool.poetry]
name = "hexlet-code"
packages = [{ include = "gendiff" },]
version = "0.1.0"
description = ""
authors = ["Alex Subach <alex_28.02@mail.ru>"]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
